Fix Coding Style  - Updated project structure once more  - Phase 3: manual fixes
