--- Adobe/AxpConnector/Observer/CheckoutOnepageControllerSuccessActionObserver.php
@@ -1,89 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-namespace Adobe\AxpConnector\Observer;
-
-use Magento\Framework\Event\ObserverInterface;
-
-/**
- * Observer for checkout success event
- */
-class CheckoutOnepageControllerSuccessActionObserver implements ObserverInterface
-{
-    /**
-     * @var string
-     */
-    const COOKIE_NAME = 'axpconnector_checkout_success';
-
-    /**
-     * Short duration, it just has to survive a page load
-     */
-    const COOKIE_DURATION_SECS = 180;
-
-    /**
-     * @var \Adobe\AxpConnector\Helper\Data
-     */
-    protected $helper;
-
-    /**
-     * @var \Psr\Log\LoggerInterface
-     */
-    protected $logger;
-
-    /**
-     * @var \Magento\Framework\Stdlib\CookieManagerInterface
-     */
-    protected $cookieManager;
-
-    /**
-     * @var \Magento\Framework\Stdlib\Cookie\CookieMetadataFactory
-     */
-    protected $cookieMetadataFactory;
-
-    /**
-     * CheckoutOnepageControllerSuccessActionObserver constructor.
-     * @param \Adobe\AxpConnector\Helper\Data $helper
-     * @param \Psr\Log\LoggerInterface $logger
-     * @param \Magento\Framework\Stdlib\CookieManagerInterface $cookieManager
-     * @param \Magento\Framework\Stdlib\Cookie\CookieMetadataFactory $cookieMetadataFactory
-     */
-    public function __construct(
-        \Adobe\AxpConnector\Helper\Data $helper,
-        \Psr\Log\LoggerInterface $logger,
-        \Magento\Framework\Stdlib\CookieManagerInterface $cookieManager,
-        \Magento\Framework\Stdlib\Cookie\CookieMetadataFactory $cookieMetadataFactory
-    ) {
-        $this->logger = $logger;
-        $this->helper = $helper;
-        $this->cookieManager = $cookieManager;
-        $this->cookieMetadataFactory = $cookieMetadataFactory;
-    }
-
-    /**
-     * @inheritdoc
-     *
-     * @param \Magento\Framework\Event\Observer $observer
-     * @throws \Magento\Framework\Exception\InputException
-     * @throws \Magento\Framework\Stdlib\Cookie\CookieSizeLimitReachedException
-     * @throws \Magento\Framework\Stdlib\Cookie\FailureToSendException
-     */
-    public function execute(\Magento\Framework\Event\Observer $observer)
-    {
-        $orderIds = $observer->getEvent()->getOrderIds();
-        if (empty($orderIds) || !is_array($orderIds)) {
-            return;
-        }
-
-        $datalayerContent = $this->helper->orderPlacedPushData($orderIds);
-
-        if (count($datalayerContent) > 0) {
-            $jsonArray = $this->helper->jsonify($datalayerContent);
-            $metadata = $this->cookieMetadataFactory
-                ->createPublicCookieMetadata()
-                ->setDuration(self::COOKIE_DURATION_SECS);
-            $this->cookieManager->setPublicCookie(self::COOKIE_NAME, $jsonArray, $metadata);
-        }
-    }
-}
