--- Adobe/AxpConnector/view/frontend/templates/checkout.phtml
@@ -1,10 +1,0 @@
-<?php
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
-?>
-<script type="text/javascript" id="axpconnector-datalayer-checkout-started">
-    window.AppEventData = window.AppEventData || [];
-    window.AppEventData.push(<?= $block->datalayerJson() ?>);
-</script>
