--- Adobe/AxpConnector/view/frontend/web/js/view/launch.js
@@ -1,13 +1,0 @@
-define([
-    'uiComponent',
-    'Magento_Customer/js/customer-data'
-], function (Component, customerData) {
-    'use strict';
-
-    return Component.extend({
-        initialize: function () {
-            this._super();
-            this.launch = customerData.get('launch');
-        }
-    });
-});
