+++ AxpConnector/Block/Display.php
@@ -1,0 +1,73 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+declare(strict_types=1);
+
+namespace Adobe\AxpConnector\Block;
+
+/**
+ * Display block.
+ *
+ * @api
+ */
+class Display extends \Magento\Framework\View\Element\Template
+{
+    /**
+     * @var array
+     */
+    protected $datalayerEvents;
+
+    /**
+     * @var \Psr\Log\LoggerInterface
+     */
+    protected $_logger;
+
+    /**
+     * @param \Magento\Framework\View\Element\Template\Context $context
+     * @param \Adobe\AxpConnector\Helper\Data $helper
+     * @param \Psr\Log\LoggerInterface $logger
+     */
+    public function __construct(
+        \Magento\Framework\View\Element\Template\Context $context,
+        \Adobe\AxpConnector\Helper\Data $helper,
+        \Psr\Log\LoggerInterface $logger
+    ) {
+        parent::__construct($context);
+        $this->helper = $helper;
+        $this->_logger = $logger;
+        $this->datalayerEvents = [];
+    }
+
+    /**
+     * Is enabled.
+     *
+     * @return bool
+     */
+    public function isEnabled()
+    {
+        return $this->helper->isEnabled();
+    }
+
+    /**
+     * Push event to datalayer.
+     *
+     * @param mixed $eventData
+     */
+    public function pushDatalayerEvent($eventData)
+    {
+        array_push($this->datalayerEvents, $eventData);
+    }
+
+    /**
+     * Encode events.
+     *
+     * @return array
+     */
+    public function encodedDatalayerEvents()
+    {
+        $encoded = array_map([$this->helper, 'jsonify'], $this->datalayerEvents);
+        return $encoded;
+    }
+}
