+++ AxpConnector/README.md
@@ -1,0 +1,10 @@
+# Hello!
+Welcome to the first Community Project that spans both Magento and Adobe Experience Platform!
+
+The Adobe Exeperience Platform Connector is a Magento module that will serve as the foundation for the subsequent Magento-Adobe integrations. Based on Launch, by Adobe, this connector will allow Magento merchants to take their businesses to the next level. 
+
+Adobe Experience Cloud offers products that help merchants become more successful. Among these products are Adobe Analytics and Adobe Target which give merchants deep customer insights and the ability to deliver personalized experiences at scale. This project is meant to simplify the implementation of Adobe Experience Cloud products down to a few clicks so merchants have better access to this tools.
+
+Here is our [ZenHub board](https://app.zenhub.com/workspace/o/magento/axp-connector/boards?repos=151743693), best viewed with the [Chrome ZenHub extension](https://chrome.google.com/webstore/detail/zenhub-for-github/ogcgkffhplmphkaahpmffcafajaocjbd). Otherwise, our backlog is organized as a list of [issues](https://github.com/magento/axp-connector/issues).
+
+To join the project, attend a regular meeting (timing TBD for now) or join [#axp-connector](https://magentocommeng.slack.com/messages/CD4EGUYEB) slack channel (sign up for EngCom Slack workspace [here](https://tinyurl.com/engcom-signup))
