+++ AxpConnector/composer.json
@@ -1,0 +1,25 @@
+{
+  "name": "adobe/module-axp-connector",
+  "require": {
+    "php": "~7.1.3||~7.2.0",
+    "magento/module-customer": "*",
+    "magento/module-checkout": "*",
+    "magento/module-catalog": "*",
+    "magento/module-quote": "*",
+    "magento/module-sales": "*",
+    "magento/module-store": "*"
+  },
+  "type": "magento2-module",
+  "license": [
+    "OSL-3.0",
+    "AFL-3.0"
+  ],
+  "autoload": {
+    "files": [
+      "registration.php"
+    ],
+    "psr-4": {
+      "Adobe\\AxpConnector\\": ""
+    }
+  }
+}
