+++ AxpConnector/etc/frontend/sections.xml
