+++ AxpConnector/etc/module.xml
