+++ AxpConnector/view/frontend/layout/catalog_product_view.xml
