+++ AxpConnector/view/frontend/layout/checkout_index_index.xml
