+++ AxpConnector/view/frontend/layout/default.xml
