+++ AxpConnector/view/frontend/requirejs-config.js
@@ -1,0 +1,11 @@
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+var config = {
+    map: {
+        '*': {
+            axpconnector: 'Adobe_AxpConnector/js/axpconnector'
+        }
+    }
+};
