+++ AxpConnector/view/frontend/templates/datalayer.phtml
@@ -1,0 +1,28 @@
+<?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+
+$_helper = $this->helper(\Adobe\AxpConnector\Helper\Data::class);
+?>
+
+<?php if ($_helper->isEnabled()) : ?>
+    <script type="text/javascript" id="axpconnector-datalayer-container">
+        // Datalayer goes here
+        window.AppEventData = window.AppEventData || [];
+        <?php foreach ($block->encodedDatalayerEvents() as $eventJson) : ?>
+        window.AppEventData.push(<?= /* @escapeNotVerified */ $eventJson ?>);
+        <?php endforeach; ?>
+    </script>
+
+
+    <script type="text/x-magento-init">
+    {
+        "*": {
+            "axpconnector": {}
+        }
+    }
+
+    </script>
+<?php endif; ?>
