+++ AxpConnector/view/frontend/templates/private_content_datalayer.phtml
