+++ AxpConnector/view/frontend/web/js/view/launch.js
@@ -1,0 +1,17 @@
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
+define([
+    'uiComponent',
+    'Magento_Customer/js/customer-data'
+], function (Component, customerData) {
+    'use strict';
+
+    return Component.extend({
+        initialize: function () {
+            this._super();
+            this.launch = customerData.get('launch');
+        }
+    });
+});
